+++ src/app/kit/timing/interfaces/duration.interface.ts
@@ -1,0 +1,10 @@
+export interface DurationInterface {
+	years?: number;
+	months?: number;
+	weeks?: number;
+	days?: number;
+	hours?: number;
+	minutes?: number;
+	seconds?: number;
+	milliseconds?: number;
+}