+++ src/app/kit/timing/utility/date.utility.ts
@@ -1,6 +1,7 @@
 import * as momentNs from 'moment';
 import { DateModel } from '../models/date.model';
 import * as _moment from 'moment-timezone';
+import { DurationInterface } from '../interfaces/duration.interface';
 const moment = _moment;
 export class DateUtility {
 	static TODAY(): DateModel {
@@ -80,4 +81,148 @@ export class DateUtility {
 			.filter((v, i) => v !== '00' || i > 0)
 			.join(':');
 	};
+
+	static isBisestile(year: number) {
+		if (year > 1584 && (year % 400 === 0 || (year % 4 === 0 && year % 100 !== 0))) {
+			return true;
+		} else {
+			return false;
+		}
+	}
+	static getMaxDayByMonth(month: number, year?: number) {
+		switch (month) {
+			case 4:
+			case 6:
+			case 9:
+			case 11:
+				return 30;
+			case 2:
+				return !year ? 28 : this.isBisestile(year) ? 29 : 28;
+			case 1:
+			case 3:
+			case 5:
+			case 7:
+			case 8:
+			case 10:
+			case 12:
+				return 31;
+			default:
+				return undefined;
+		}
+	}
+
+	// operations - add
+	static add(
+		date: any | Date | string | number | DateModel,
+		duration: DurationInterface,
+		format?: string,
+	) {
+		const dta = new DateModel(date, format);
+		return new DateModel(dta.date.clone().add(duration), format);
+	}
+	static addYears(date: any | Date | string | number | DateModel, years: number, format?: string) {
+		return this.add(date, { years: years }, format);
+	}
+	static addMonths(
+		date: any | Date | string | number | DateModel,
+		months: number,
+		format?: string,
+	) {
+		return this.add(date, { months: months }, format);
+	}
+	static addWeeks(date: any | Date | string | number | DateModel, weeks: number, format?: string) {
+		return this.add(date, { weeks: weeks }, format);
+	}
+	static addDays(date: any | Date | string | number | DateModel, days: number, format?: string) {
+		return this.add(date, { days: days }, format);
+	}
+	static addHours(date: any | Date | string | number | DateModel, hours: number, format?: string) {
+		return this.add(date, { hours: hours }, format);
+	}
+	static addMinutes(
+		date: any | Date | string | number | DateModel,
+		minutes: number,
+		format?: string,
+	) {
+		return this.add(date, { minutes: minutes }, format);
+	}
+	static addSeconds(
+		date: any | Date | string | number | DateModel,
+		seconds: number,
+		format?: string,
+	) {
+		return this.add(date, { seconds: seconds }, format);
+	}
+	static addMilliseconds(
+		date: any | Date | string | number | DateModel,
+		milliseconds: number,
+		format?: string,
+	) {
+		return this.add(date, { milliseconds: milliseconds }, format);
+	}
+	// operations - subtract
+	static subtract(
+		date: any | Date | string | number | DateModel,
+		duration: DurationInterface,
+		format?: string,
+	) {
+		const dta = new DateModel(date, format);
+		return new DateModel(dta.date.clone().subtract(duration), format);
+	}
+	static subtractYears(
+		date: any | Date | string | number | DateModel,
+		years: number,
+		format?: string,
+	) {
+		return this.subtract(date, { years: years }, format);
+	}
+	static subtractMonths(
+		date: any | Date | string | number | DateModel,
+		months: number,
+		format?: string,
+	) {
+		return this.subtract(date, { months: months }, format);
+	}
+	static subtractWeeks(
+		date: any | Date | string | number | DateModel,
+		weeks: number,
+		format?: string,
+	) {
+		return this.subtract(date, { weeks: weeks }, format);
+	}
+	static subtractDays(
+		date: any | Date | string | number | DateModel,
+		days: number,
+		format?: string,
+	) {
+		return this.subtract(date, { days: days }, format);
+	}
+	static subtractHours(
+		date: any | Date | string | number | DateModel,
+		hours: number,
+		format?: string,
+	) {
+		return this.subtract(date, { hours: hours }, format);
+	}
+	static subtractMinutes(
+		date: any | Date | string | number | DateModel,
+		minutes: number,
+		format?: string,
+	) {
+		return this.subtract(date, { minutes: minutes }, format);
+	}
+	static subtractSeconds(
+		date: any | Date | string | number | DateModel,
+		seconds: number,
+		format?: string,
+	) {
+		return this.subtract(date, { seconds: seconds }, format);
+	}
+	static subtractMilliseconds(
+		date: any | Date | string | number | DateModel,
+		milliseconds: number,
+		format?: string,
+	) {
+		return this.subtract(date, { milliseconds: milliseconds }, format);
+	}
 }