+++ src/app/modules/shop-payment/dtos/balance-flow.filter.ts
@@ -2,6 +2,10 @@ export class BalanceFlowFilter {
 	public dtaFrom: string;
 	public dtaTo: string;
 	public groups: string[];
+	public causal: string;
 	public limit: number;
 	public page: number;
+	public flgPayments: number;
+	public flgGroupcod: number;
+	public year: number;
 }