+++ src/app/shared/models/pagination/col-table.model.ts
@@ -4,4 +4,5 @@ export class ColTableModel {
 	label: string | StringTranslate;
 	key: string;
 	fieldModel: string;
+	template?: any;
 }