+++ src/app/kit/CHANGELOG.md
@@ -1,12 +1,17 @@
 # CHANGELOG KIT
 
-## 2.0.1 - XX/03/2022
+## 2.1.0 - XX/03/2022
 
 > Versione che 
 
 ### changes
 
 - `[fix]` Aggiunto il controllo sull'esistenza dei campi in input nel metodo `getBase64ByContent` di `FileService` del modulo `filemanager`
+- `[feature]` **locale**: Aggiunta la pipe `showcurrency`
+- `[feature]` **timing**: Aggiunta l'interfaccia `DurationInterface` e tutti i metodi di utility in `DateUtility` per aggiungere o sottrarre valori ad una data
+- `[feature]` **timing**: Aggiunti in `DateUtility` le funzioni `isBisestile`, per calcolare se un anno è bisestile, e `getMaxDayByMonth` per calcolare il numero di giorni di un mese
+- `[feature]` **config**: Aggiunta in `ArrayUtility` la funzione `sum`, per calcolare la somma di una property presente in array
+- `[feature]` **charts**: Aggiunti in `ChartComponent` l'input `appendText`, per aggiungere un testo ad un valore in tooltip
 
 ## 2.0.0 - 14/03/2022